Added TODO tasks
The deadline is Sunday, so do your best to work on as many of these tasks as possible, in the order they are laid out (that's important). It doesn't really matter to me if you don't finish all of them, or even most of them, by the arbitrary deadline so long as you do decent work with the tasks you managed to work on.